requirements.txt
Django==4.2.28
python-dotenv==1.0.1
psycopg2-binary==2.9.9
celery==5.4.0
requests==2.32.3
djangorestframework==3.15.2
redis==5.0.8
